+++ app/src/main/java/com/example/playlistmaker/recyclerView/TrackAdapter.kt
@@ -1,8 +1,9 @@
-package com.example.playlistmaker
+package com.example.playlistmaker.recyclerView
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
+import com.example.playlistmaker.R
 
 class TrackAdapter(private val tracks: List<Track>) : RecyclerView.Adapter<TrackViewHolder>() {
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): TrackViewHolder {
