+++ gradle/libs.versions.toml
@@ -1,5 +1,5 @@
 [versions]
-agp = "8.4.2"
+agp = "8.5.0"
 glide = "4.14.2"
 kotlin = "1.9.0"
 coreKtx = "1.13.1"
